@@ -13,21 +13,22 @@
     "@plone/blocks": "workspace: *",
     "@plone/client": "workspace: *",
     "@plone/components": "workspace: *",
-    "@plone/registry": "workspace: *",
     "@plone/providers": "workspace: *",
-    "@tanstack/react-query": "^5.37.1",
-    "next": "14.2.2",
+    "@plone/registry": "workspace: *",
+    "@tanstack/react-query": "^5.59.0",
+    "next": "14.2.14",
     "react": "^18",
-    "react-aria-components": "^1.1.1",
+    "react-aria-components": "^1.4.0",
     "react-dom": "^18"
   },
   "devDependencies": {
-    "@tanstack/react-query-devtools": "^5.37.1",
+    "@plone/types": "workspace: *",
+    "@tanstack/react-query-devtools": "^5.59.0",
     "@types/node": "^20",
     "@types/react": "^18",
     "@types/react-dom": "^18",
     "eslint": "^8",
-    "eslint-config-next": "14.2.2",
-    "typescript": "^5.4.5"
+    "eslint-config-next": "14.2.14",
+    "typescript": "^5.6.3"
   }
 }
