@@ -348,7 +348,7 @@
     "ts-jest": "^26.4.2",
     "ts-loader": "9.4.4",
     "tsconfig": "*",
-    "typescript": "^5.4.2",
+    "typescript": "^5.6.3",
     "use-trace-update": "1.3.2",
     "wait-on": "6.0.0",
     "webpack": "5.90.1",
