@@ -27,7 +27,7 @@
     "@tanstack/react-query-devtools": "^5.37.1",
     "@types/react": "^18.2.20",
     "@types/react-dom": "^18.2.7",
-    "typescript": "^5.4.2",
+    "typescript": "^5.6.3",
     "vite": "^5.1.0",
     "vite-tsconfig-paths": "^4.2.1"
   },
