@@ -25,7 +25,7 @@
     "axios": "^1.6.5",
     "get-port": "^7.0.0",
     "react": "^18.2.0",
-    "react-aria-components": "^1.1.1",
+    "react-aria-components": "^1.4.0",
     "react-dom": "^18.2.0",
     "sirv": "^2.0.4"
   },
@@ -44,7 +44,7 @@
     "isbot": "^4.3.0",
     "node-fetch": "^3.3.2",
     "serve-static": "^1.15.0",
-    "typescript": "^5.4.2",
+    "typescript": "^5.6.3",
     "vite": "^5.2.9",
     "vite-plugin-babel": "^1.2.0"
   }
