@@ -77,7 +77,7 @@
     "parcel": "^2.12.0",
     "release-it": "17.1.1",
     "tsconfig": "workspace:*",
-    "typescript": "5.2.2",
-    "vitest": "^1.3.1"
+    "typescript": "^5.4.5",
+    "vitest": "^2.1.3"
   }
 }
