@@ -56,6 +56,6 @@
     "react-intl": "3.12.1",
     "release-it": "^17.1.1",
     "tsconfig": "workspace:*",
-    "typescript": "5.4.2"
+    "typescript": "^5.6.3"
   }
 }
